requirements.txt
textual==0.27.0
PyYAML==6.0
psutil==5.9.5
rich==13.3.5

